backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Computer Vision and AI
opencv-python==4.8.1.78
mediapipe==0.10.8
tensorflow==2.15.0
torch==2.1.1
torchvision==0.16.1
numpy==1.24.3
pillow==10.1.0

# Image Processing
scikit-image==0.22.0
imutils==0.5.4

# Async and Concurrency
asyncio==3.4.3
aiofiles==23.2.1
asyncpg==0.29.0

# Security
cryptography==41.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# WebSocket
websockets==12.0

# Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0
psutil==5.9.6
fluent-logger==0.10.0
elasticsearch==8.10.0
python-json-logger==2.0.7

# Utilities
python-multipart==0.0.6
email-validator==2.1.0
python-dateutil==2.8.2

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8